Fix a few compiler warnings (#60)
* Fix compiler warnings.

* Use c++ style casting instead of c style casting.

+++ src/GUI/Widgets/LineEdit.cpp
@@ -18,7 +18,7 @@ LineEdit::~LineEdit()
 
 void LineEdit::Init()
 {
-    cp = line.size();
+    cp = static_cast<int>(line.size());
 }
 
 void LineEdit::BeginInput()
